Add uefi bios vars, exec cmd line and log to file parameters

+++ cmd/manager/script/launch-qemu.sh
@@ -3,16 +3,18 @@
 #
 # user changeable parameters
 #
-HDA_FILE="${HOME}/ubuntu-18.04-desktop.qcow2"
+
+HDA_FILE="cmd/manager/img/focal-server-cloudimg-amd64.qcow2"
 GUEST_SIZE_IN_MB="4096"
 SEV_GUEST="1"
 SMP_NCPUS="4"
-CONSOLE="qxl"
-QEMU_INSTALL_DIR=/usr/local/bin/
-UEFI_BIOS_CODE="/usr/local/share/qemu/OVMF_CODE.fd"
-UEFI_BIOS_VARS="OVMF_VARS.fd"
-#VNC_PORT=""
+CONSOLE="serial" # for gtk, use qxl
+VNC_PORT=""
 USE_VIRTIO="1"
+
+UEFI_BIOS_CODE="/usr/share/OVMF/OVMF_CODE.fd"
+UEFI_BIOS_VARS_ORIG="/usr/share/OVMF/OVMF_VARS.fd"
+UEFI_BIOS_VARS_COPY="cmd/manager/img/OVMF_VARS.fd"
 
 CBITPOS=51
 HOST_HTTP_PORT=9301
@@ -20,49 +22,39 @@ GUEST_HTTP_PORT=9031
 HOST_GRPC_PORT=7020
 GUEST_GRPC_PORT=7002
 
+ENABLE_FILE_LOG="0"
+EXEC_QEMU_CMDLINE="0"
+
 usage() {
     echo "$0 [options]"
     echo "Available <commands>:"
-    echo " -hda          hard disk ($HDA_FILE)"
-    echo " -nosev        disable sev support"
-    echo " -mem          guest memory"
-    echo " -smp          number of cpus"
-    echo " -console      display console to use (serial or gxl)"
-    echo " -vnc          VNC port to use"
-    echo " -bios         bios to use (default $UEFI_BIOS_CODE)"
-    echo " -kernel       kernel to use"
-    echo " -initrd       initrd to use"
-    echo " -cdrom        CDROM image"
-    echo " -virtio       use virtio devices"
-    echo " -gdb          start gdbserver"
-    echo " -cbitpos      location of the C-bit"
-    echo " -hosthttp     host http port"
-    echo " -guesthttp    guest http port"
-    echo " -hostgrpc     host grpc port"
-    echo " -guestgrpc    guest grpc port"
+    echo " -hda           hard disk ($HDA_FILE)"
+    echo " -nosev         disable sev support"
+    echo " -mem           guest memory"
+    echo " -smp           number of cpus"
+    echo " -console       display console to use (serial or gxl)"
+    echo " -vnc           VNC port to use"
+    echo " -bios          bios to use (default $UEFI_BIOS_CODE)"
+    echo " -kernel        kernel to use"
+    echo " -initrd        initrd to use"
+    echo " -cdrom         CDROM image"
+    echo " -virtio        use virtio devices"
+    echo " -cbitpos       location of the C-bit"
+    echo " -hosthttp      host http port"
+    echo " -guesthttp     guest http port"
+    echo " -hostgrpc      host grpc port"
+    echo " -guestgrpc     guest grpc port"
+    echo " -origuefivars  UEFI BIOS vars original file (default $UEFI_BIOS_VARS_ORIG)"
+    echo " -copyuefivars  UEFI BIOS vars copy file (default $UEFI_BIOS_VARS_COPY)"
+    echo " -exec          execute the QEMU command (default $EXEC_QEMU_CMDLINE)"
+    echo " -filelog       enable/disable QEMU cmd line file log (default: $ENABLE_FILE_LOG)"
     exit 1
 }
-
-add_opts() {
-    echo -n "$* " >> ${QEMU_CMDLINE}
-}
-
-run_cmd() {
-    if ! "$@"; then
-        echo "Command '$*' failed"
-        exit 1
-    fi
-}
-
-if [ "$(id -u)" -ne 0 ]; then
-    echo "This script must be run as root!"
-    exit 1
-fi
 
 while [[ $1 != "" ]]; do
     case "$1" in
         -hda)
-            HDA_FILE="${2}"
+            HDA_FILE=${2}
             shift
             ;;
         -nosev)
@@ -83,16 +75,9 @@ while [[ $1 != "" ]]; do
         -vnc)
             VNC_PORT=$2
             shift
-            if [ "${VNC_PORT}" = "" ]; then
-                usage
-            fi
             ;;
         -bios)
-            UEFI_BIOS_CODE=$(readlink -f "$2")
-            shift
-            ;;
-        -netconsole)
-            NETCONSOLE_PORT=$2
+            UEFI_BIOS_CODE=$2
             shift
             ;;
         -initrd)
@@ -110,35 +95,74 @@ while [[ $1 != "" ]]; do
         -virtio)
             USE_VIRTIO="1"
             ;;
-        -gdb)
-            USE_GDB="1"
-            ;;
         -cbitpos)
             CBITPOS=$2
+            shift
             ;;
         -hosthttp)
             HOST_HTTP_PORT=$2
+            shift
             ;;
         -guesthttp)
             GUEST_HTTP_PORT=$2
+            shift
             ;;
         -hostgrpc)
             HOST_GRPC_PORT=$2
+            shift
             ;;
-        -guestgrpc)
-            GUEST_GRPC_PORT=$2
+        -origuefivars)
+            UEFI_BIOS_VARS_ORIG=$2
+            shift
             ;;
+        -copyuefivars)
+            UEFI_BIOS_VARS_COPY=$2
+            shift
+            ;;        
+        -exec)
+            EXEC_QEMU_CMDLINE="1"
+            ;;      
+        -filelog)
+            ENABLE_FILE_LOG="1"
+            ;;                           
         *)
             usage;;
     esac
     shift
 done
 
+#
+# func definitions
+#
+
+add_opts() {
+    echo -n "$* " >> ${QEMU_CMDLINE}
+}
+
+run_cmd() {
+    if ! "$@"; then
+        echo "Command '$*' failed"
+        exit 1
+    fi
+}
+
+if [ "$(id -u)" -ne 0 ]; then
+    echo "This script must be run as root!"
+    exit 1
+fi
+
+# copy BIOS variables to new dest for VM use without modifying the original ones
+cp "$UEFI_BIOS_VARS_ORIG" "$UEFI_BIOS_VARS_COPY"
+
+#
+# Qemu cmd line construction
+#
+
 # we add all the qemu command line options into a file
 QEMU_CMDLINE=/tmp/cmdline.$$
 rm -rf ${QEMU_CMDLINE}
 
-add_opts "${QEMU_INSTALL_DIR}qemu-system-x86_64"
+add_opts "$(which qemu-system-x86_64)"
 
 # Basic virtual machine property
 add_opts "-enable-kvm -cpu EPYC -machine q35"
@@ -152,8 +176,8 @@ add_opts "-m ${GUEST_SIZE_IN_MB}M,slots=5,maxmem=30G"
 # The OVMF binary, including the non-volatile variable store, appears as a
 # "normal" qemu drive on the host side, and it is exposed to the guest as a
 # persistent flash device.
-add_opts "-drive if=pflash,format=raw,unit=0,file=${UEFI_BIOS_CODE},readonly"
-add_opts "-drive if=pflash,format=raw,unit=1,file=${UEFI_BIOS_VARS}"
+add_opts "-drive if=pflash,format=raw,unit=0,file=${UEFI_BIOS_CODE},readonly=on"
+add_opts "-drive if=pflash,format=raw,unit=1,file=${UEFI_BIOS_VARS_COPY}"
 
 # add CDROM if specified
 [ -n "$CDROM_FILE" ] && add_opts "-drive file=${CDROM_FILE},media=cdrom -boot d"
@@ -206,22 +230,43 @@ fi
 # start monitor on pty
 add_opts "-monitor pty"
 
-# log the console  output in stdout.log
-QEMU_CONSOLE_LOG=$(pwd)/stdout.log
+#
+# Qemu cmd line log
+#
 
-# save the command line args into log file
-cat $QEMU_CMDLINE | tee "${QEMU_CONSOLE_LOG}"
-echo | tee -a "${QEMU_CONSOLE_LOG}"
+# Set the log file path if ENABLE_FILE_LOG is 1
+if [ "$ENABLE_FILE_LOG" = "1" ]; then
+    LOG_FILE=$(pwd)/stdout.log
 
+    # Save the command line args into log file
+    cat "$QEMU_CMDLINE" > "$LOG_FILE"
+    echo >> "$LOG_FILE"
+fi
+
+ # Log the command line to the console
+cat "$QEMU_CMDLINE"
+echo
+
+#
+# Qemu cmd line execution
+#
+
+if [[ "${EXEC_QEMU_CMDLINE}" = "0" ]]; then
+    exit 0
+fi
 
 # map CTRL-C to CTRL ]
 echo "Mapping CTRL-C to CTRL-]"
 stty intr ^]
 
-            echo "Launching VM ..."
-            bash ${QEMU_CMDLINE} 2>&1 | tee -a "${QEMU_CONSOLE_LOG}"
+echo "Launching VM ..."
+if [ "$ENABLE_FILE_LOG" = "1" ]; then
+    bash ${QEMU_CMDLINE} 2>&1 | tee -a "${LOG_FILE}"
+else 
+    bash ${QEMU_CMDLINE} 2>&1
+fi
 
-            # restore the mapping
-            stty intr ^c
+# restore the mapping
+stty intr ^c
 
-            rm -rf ${QEMU_CMDLINE}
+rm -rf ${QEMU_CMDLINE}
